refactor: jouranlsPhotos, journalsDogs 관련 메소드 해당 모듈로 이전
